implementado validacao pt1

+++ app/Http/Controllers/MarcaController.php
@@ -46,6 +46,10 @@ class MarcaController extends Controller
     public function show($id)
     {
         $marca = $this->marca->find($id);
+        if($marca === null){
+            return ['erro' => 'Recurso pesquisado não existe'];
+        }
+
         return $marca;
     }
 
@@ -63,7 +67,12 @@ class MarcaController extends Controller
     public function update(Request $request, $id)
     {
         //$marca->update($request->all());
+
         $marca = $this->marca->find($id);
+        if($marca === null){
+            return ['erro' => 'Impossível realizar a atualização. O recurso solicitado não existe'];
+        }
+
         $marca->update($request->all());
         return $marca;
     }
@@ -74,6 +83,10 @@ class MarcaController extends Controller
     public function destroy($id)
     {
         $marca = $this->marca->find($id);
+
+        if($marca === null){
+            return ['erro' => 'Impossível realizar a exclusão. O recurso solicitado não existe'];
+        }
         $marca->delete();
         return ['msg' => 'A marca foi removida com sucesso'];
     }
